Pull SQL out into string const.
